Combine tests, fix generation, eager initialize baseScope
